cap5_deteccion_pronostico/requirements.txt
pandas
numpy
scipy
statsmodels
scikit-learn
matplotlib
seaborn
graphviz
Pillow
